Implemented FlattenLens for MySQL
