+++ test/lightweight-tests/src/test/resources/nested/mysql/nested-mysql.properties
@@ -1,0 +1,4 @@
+jdbc.url = jdbc:mysql://${docker.url}:3694/nested
+jdbc.user = root
+jdbc.password = ${docker.mysql.password}
+jdbc.driver = com.mysql.cj.jdbc.Driver
